new-relic-exporter/requirements.txt
quixstreams==2.2.0a0
python-dotenv
aiohttp